requirements.txt
streamlit
pandas
numpy
matplotlib
sympy
scipy.interpolate
